trainmodel-demo/requirements.txt
bandit-cli==0.0.4
pandas==0.19.1
statsmodels==0.6.1
